Move retrieveQuestionResponseFromApplication()
We move the function from utils to applicationUtils. We will want to use assertUnreachable() in our RestClient class, and without moving this function, we would create a circular dependency

Update applicationUtils.test.ts
